// src/content/perks.generated.ts
export type PerkRefId = number;

export type PerkId =
  | 'antiperk'
  | 'bloody_mess_quick_learner'
  | 'sharpshooter'
  | 'fastloader'
  | 'lean_mean_exp_machine'
  | 'long_distance_runner'
  | 'pyrokinetic'
  | 'instant_winner'
  | 'grim_deal'
  | 'alternate_weapon'
  | 'plaguebearer'
  | 'evil_eyes'
  | 'ammo_maniac'
  | 'radioactive'
  | 'fastshot'
  | 'fatal_lottery'
  | 'random_weapon'
  | 'mr_melee'
  | 'anxious_loader'
  | 'final_revenge'
  | 'telekinetic'
  | 'perk_expert'
  | 'unstoppable'
  | 'regression_bullets'
  | 'infernal_contract'
  | 'poison_bullets'
  | 'dodger'
  | 'bonus_magnet'
  | 'uranium_filled_bullets'
  | 'doctor'
  | 'monster_vision'
  | 'hot_tempered'
  | 'bonus_economist'
  | 'thick_skinned'
  | 'barrel_greaser'
  | 'ammunition_within'
  | 'veins_of_poison'
  | 'toxic_avenger'
  | 'regeneration'
  | 'pyromaniac'
  | 'ninja'
  | 'highlander'
  | 'jinxed'
  | 'perk_master'
  | 'reflex_boosted'
  | 'greater_regeneration'
  | 'breathing_room'
  | 'death_clock'
  | 'my_favourite_weapon'
  | 'bandage'
  | 'angry_reloader'
  | 'ion_gun_master'
  | 'stationary_reloader'
  | 'man_bomb'
  | 'fire_caugh'
  | 'living_fortress'
  | 'tough_reloader'
  | 'lifeline_50_50';

export interface PerkDef {
  id: PerkId;
  refId: PerkRefId;
  name: string;
  description: string;
  maxStacks: number;
  rarity: 'common' | 'uncommon' | 'rare' | 'legendary';
  prereq?: PerkId;
  tags?: string[];
}
// ...
export const PERK_BY_ID: Record<PerkId, PerkDef> = {
  'antiperk': PERKS.find((p) => p.id === 'antiperk')!,
  'bloody_mess_quick_learner': PERKS.find((p) => p.id === 'bloody_mess_quick_learner')!,
  'sharpshooter': PERKS.find((p) => p.id === 'sharpshooter')!,
  'fastloader': PERKS.find((p) => p.id === 'fastloader')!,
  'lean_mean_exp_machine': PERKS.find((p) => p.id === 'lean_mean_exp_machine')!,
  'long_distance_runner': PERKS.find((p) => p.id === 'long_distance_runner')!,
  'pyrokinetic': PERKS.find((p) => p.id === 'pyrokinetic')!,
  'instant_winner': PERKS.find((p) => p.id === 'instant_winner')!,
  'grim_deal': PERKS.find((p) => p.id === 'grim_deal')!,
  'alternate_weapon': PERKS.find((p) => p.id === 'alternate_weapon')!,
  'plaguebearer': PERKS.find((p) => p.id === 'plaguebearer')!,
  'evil_eyes': PERKS.find((p) => p.id === 'evil_eyes')!,
  'ammo_maniac': PERKS.find((p) => p.id === 'ammo_maniac')!,
  'radioactive': PERKS.find((p) => p.id === 'radioactive')!,
  'fastshot': PERKS.find((p) => p.id === 'fastshot')!,
  'fatal_lottery': PERKS.find((p) => p.id === 'fatal_lottery')!,
  'random_weapon': PERKS.find((p) => p.id === 'random_weapon')!,
  'mr_melee': PERKS.find((p) => p.id === 'mr_melee')!,
  'anxious_loader': PERKS.find((p) => p.id === 'anxious_loader')!,
  'final_revenge': PERKS.find((p) => p.id === 'final_revenge')!,
  'telekinetic': PERKS.find((p) => p.id === 'telekinetic')!,
  'perk_expert': PERKS.find((p) => p.id === 'perk_expert')!,
  'unstoppable': PERKS.find((p) => p.id === 'unstoppable')!,
  'regression_bullets': PERKS.find((p) => p.id === 'regression_bullets')!,
  'infernal_contract': PERKS.find((p) => p.id === 'infernal_contract')!,
  'poison_bullets': PERKS.find((p) => p.id === 'poison_bullets')!,
  'dodger': PERKS.find((p) => p.id === 'dodger')!,
  'bonus_magnet': PERKS.find((p) => p.id === 'bonus_magnet')!,
  'uranium_filled_bullets': PERKS.find((p) => p.id === 'uranium_filled_bullets')!,
  'doctor': PERKS.find((p) => p.id === 'doctor')!,
  'monster_vision': PERKS.find((p) => p.id === 'monster_vision')!,
  'hot_tempered': PERKS.find((p) => p.id === 'hot_tempered')!,
  'bonus_economist': PERKS.find((p) => p.id === 'bonus_economist')!,
  'thick_skinned': PERKS.find((p) => p.id === 'thick_skinned')!,
  'barrel_greaser': PERKS.find((p) => p.id === 'barrel_greaser')!,
  'ammunition_within': PERKS.find((p) => p.id === 'ammunition_within')!,
  'veins_of_poison': PERKS.find((p) => p.id === 'veins_of_poison')!,
  'toxic_avenger': PERKS.find((p) => p.id === 'toxic_avenger')!,
  'regeneration': PERKS.find((p) => p.id === 'regeneration')!,
  'pyromaniac': PERKS.find((p) => p.id === 'pyromaniac')!,
  'ninja': PERKS.find((p) => p.id === 'ninja')!,
  'highlander': PERKS.find((p) => p.id === 'highlander')!,
  'jinxed': PERKS.find((p) => p.id === 'jinxed')!,
  'perk_master': PERKS.find((p) => p.id === 'perk_master')!,
  'reflex_boosted': PERKS.find((p) => p.id === 'reflex_boosted')!,
  'greater_regeneration': PERKS.find((p) => p.id === 'greater_regeneration')!,
  'breathing_room': PERKS.find((p) => p.id === 'breathing_room')!,
  'death_clock': PERKS.find((p) => p.id === 'death_clock')!,
  'my_favourite_weapon': PERKS.find((p) => p.id === 'my_favourite_weapon')!,
  'bandage': PERKS.find((p) => p.id === 'bandage')!,
  'angry_reloader': PERKS.find((p) => p.id === 'angry_reloader')!,
  'ion_gun_master': PERKS.find((p) => p.id === 'ion_gun_master')!,
  'stationary_reloader': PERKS.find((p) => p.id === 'stationary_reloader')!,
  'man_bomb': PERKS.find((p) => p.id === 'man_bomb')!,
  'fire_caugh': PERKS.find((p) => p.id === 'fire_caugh')!,
  'living_fortress': PERKS.find((p) => p.id === 'living_fortress')!,
  'tough_reloader': PERKS.find((p) => p.id === 'tough_reloader')!,
  'lifeline_50_50': PERKS.find((p) => p.id === 'lifeline_50_50')!,
};